Group references' descriptions get saved
