+++ src/pages/User/EditUserProfilePage.tsx
@@ -6,38 +6,14 @@ import { Loading } from '../../components/core';
 import { useApolloErrorHandler, useNotification, useUpdateNavigation, useUrlParams, useUserContext } from '../../hooks';
 import { useCreateTagsetOnProfileMutation, useUpdateUserMutation, useUserQuery } from '../../hooks/generated/graphql';
 import { EditMode } from '../../models/editMode';
-import { UpdateUserInput, User } from '../../models/graphql-schema';
+import { User } from '../../models/graphql-schema';
 import { UserModel } from '../../models/User';
 import { logger } from '../../services/logging/winston/logger';
 import { buildUserProfileUrl } from '../../utils/urlBuilders';
 import { PageProps } from '../common';
+import { getUpdateUserInput } from '../../utils/getUpdateUserInput';
 
 interface EditUserProfilePageProps extends PageProps {}
-
-type HavingId<Entity extends { id?: unknown }> = Entity & { id: Exclude<Entity['id'], undefined> };
-const doesHaveId = <Entity extends { id?: unknown }>(entity: Entity): entity is HavingId<Entity> => !!entity.id;
-const convertIdAttrToUppercase = <Id extends string, Entity extends { id: Id }>({ id, ...attrs }: Entity) => {
-  return {
-    ID: id,
-    ...attrs,
-  };
-};
-
-// TODO [ATS] Need optimization this code is copy paste a few times.
-export const getUpdateUserInput = (user: UserModel): UpdateUserInput => {
-  const { id: userID, email, memberof, profile, agent, ...rest } = user;
-
-  return {
-    ...rest,
-    ID: userID,
-    profileData: {
-      ID: user.profile.id || '',
-      description: profile.description,
-      references: profile.references.filter(doesHaveId).map(convertIdAttrToUppercase),
-      tagsets: profile.tagsets.filter(doesHaveId).map(convertIdAttrToUppercase),
-    },
-  };
-};
 
 export const EditUserProfilePage: FC<EditUserProfilePageProps> = ({ paths }) => {
   const navigate = useNavigate();
